@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "wtforglib"
 description = "A set of utility functions I use in my various python projects."
-version = "0.8.3-dev0"
+version = "1.0.0-dev0"
 license = "MIT"
 
 authors = [
@@ -10,58 +10,64 @@ authors = [
 
 readme = "README.md"
 
-
 homepage = "https://github.com/wtfo-guru/wtforglib"
 documentation = "https://wtforglib.readthedocs.io/en/stable"
 
 keywords = []
 
 classifiers = [
-  "Development Status :: 3 - Alpha",
-  "Programming Language :: Python :: 3.7",
-  "Programming Language :: Python :: 3.8",
+  "Development Status :: 4 - Beta",
+  "License :: OSI Approved :: MIT License",
+  "Programming Language :: Python :: 3 :: Only",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
   "Intended Audience :: Developers",
   "Operating System :: OS Independent",
   "Topic :: Software Development :: Libraries :: Python Modules",
   "Typing :: Typed",
 ]
 
+packages = [
+    { include = "wtforglib" },
+    { include = "wtforglib/py.typed" },
+]
+
 [tool.poetry.dependencies]
-python = "^3.7"
-types-pyyaml = "^6.0.12.2"
-wheel = ">= 0.37.1"
+python = ">=3.9.0,<4.0"
+types-pyyaml = "^6.0.12"
 pyyaml = "^6.0"
 jinja2 = "^3.1.2"
 
+[tool.poetry.group.test]
+optional = true
+
 [tool.poetry.group.test.dependencies]
-mypy = ">=0.982,<1.2"
-
-wemake-python-styleguide = "^0.17"
-flake8-pytest-style = "^1.6"
+mypy = "^1.9"
+wemake-python-styleguide = "^0.19"
+flake8-pytest-style = ">=1.6,<3.0"
 doc8 = "^1.0"
-nitpick = "^0.32"
-
-pytest = "^7.1"
-pytest-cov = "^4.0"
+nitpick = "^0.35"
+pytest = ">=7.4,<9.0"
+pytest-cov = ">=4,<6"
 pytest-randomly = "^3.12"
-black = ">=22.12,<24.0"
-setuptools = ">=65.6.3,<68.0.0"
-packaging = ">=22,<24"
-isort = { version = "<5.12.0", python = "<3.8" }
-importlib-metadata = { version = "<5.0", python = "<3.8" }
-testfixtures = "^7.0.4"
+black = ">=23.12,<25.0"
+packaging = ">=22,<25"
+isort = "^5.13"
+testfixtures = ">=7.0.4,<9.0.0"
 pyfakefs = "^5.1.0"
-requests-cache = "^0.9"
+requests-cache = ">=0.9,<1.3"
+importlib-metadata = "^7.0"
+flake8 = ">=6.1,<8.0"
+safety = ">=2.3,<4.0"
 
 [tool.poetry.group.docs]
 optional = true
 
 [tool.poetry.group.docs.dependencies]
-sphinx = "^5.2"
-sphinx-autodoc-typehints = "^1.19"
+sphinx = ">=5.2,<8.0"
+sphinx-autodoc-typehints = ">=1.24,<3.0"
 m2r2 = "^0.3"
 tomli = "^2.0"
 
