@@ -1,9 +1,13 @@
 """Top-level module for wtforglib Library."""
+
 import platform
+import re
 import subprocess  # noqa: S404
-from typing import ClassVar
+from pathlib import Path
+from typing import ClassVar, Optional
 
 from wtforglib.errors import ShellError
+from wtforglib.kinds import FileName
 
 WINDOZE = "Windows"
 
@@ -13,7 +17,7 @@ if platform.system() == WINDOZE:
 WINDOZE_NOT_IMPLEMENTED = "{0} is not implemented on Windows"
 
 
-class WtfSingleton(object):
+class WtfSingleton:
     """Singleton class."""
 
     dn: ClassVar[str] = ""
@@ -27,7 +31,7 @@ class WtfSingleton(object):
 
 
 def domainname(test: bool = False) -> str:  # noqa: WPS605, WPS231
-    """Retuns hosts domainname.
+    """Return hosts domain name.
 
     Parameters
     ----------
@@ -37,12 +41,12 @@ def domainname(test: bool = False) -> str:  # noqa: WPS605, WPS231
     Returns
     -------
     str
-        domainname
+        domain name
 
     Raises
     ------
     ShellError
-        if subprocess returncode is not 0
+        if subprocess return code is not 0
     """
     if not WtfSingleton.dn:
         if test:
@@ -60,13 +64,13 @@ def domainname(test: bool = False) -> str:  # noqa: WPS605, WPS231
                 encoding="utf8",
             )
             if sp_result.returncode != 0:  # pragma no cover
-                raise ShellError("Failed to determine host's domainname")
+                raise ShellError("Failed to determine host's domain name")
             WtfSingleton.dn = sp_result.stdout.rstrip()
     return WtfSingleton.dn
 
 
 def hostname(test: bool = False) -> str:  # noqa: WPS605, WPS231
-    """Retuns hosts hostname.
+    """Return hosts' hostname.
 
     Parameters
     ----------
@@ -81,7 +85,7 @@ def hostname(test: bool = False) -> str:  # noqa: WPS605, WPS231
     Raises
     ------
     ShellError
-        if subprocess returncode is not 0
+        if subprocess return code is not 0
     """
     if not WtfSingleton.hn:
         if test:
@@ -126,3 +130,24 @@ def windoze_not_implemented(foo_name: str) -> None:
         raise NotImplementedError(
             WINDOZE_NOT_IMPLEMENTED.format(foo_name),
         )  # pragma: no cover
+
+
+def unix_basename(fn: FileName, ext: Optional[str] = None) -> str:
+    """Return unix basename of file name.
+
+    Parameters
+    ----------
+    fn : FileName
+        The file to get basename of
+    ext : str Optional
+        File extension
+
+    Returns
+    -------
+    str
+        unix basename of file name
+    """
+    rtn = Path(fn).name
+    if ext:
+        return re.sub("{0}$".format(ext), "", rtn)
+    return rtn
