@@ -1,14 +1,24 @@
 """Top-level module for Wtforg Library."""
 
+import logging
+import sys
+from types import MappingProxyType
 from typing import Dict, Optional, Union
-
-from wtforglib.scribe import LOG_LEVELS, Scribe
 
 OptionsDict = Dict[str, Union[str, int, bool]]
 
 OPTIONS_DEBUG = "debug"
 OPTIONS_TEST = "test"
 OPTIONS_VERBOSE = "verbose"
+LOG_LEVELS = MappingProxyType(
+    {
+        "crit": logging.CRITICAL,
+        "error": logging.ERROR,
+        "warn": logging.WARNING,
+        "info": logging.INFO,
+        "debug": logging.DEBUG,
+    },
+)
 
 
 def log_level_name(level: int, default: str = "info") -> str:
@@ -60,11 +70,48 @@ def basic_options(
     return od
 
 
-class Options(object):  # noqa: WPS214
+class SimpleScribe:
+    """Class for simple screen logging."""
+
+    _errors: int
+
+    def __init__(self) -> None:
+        """Construct a simple scribe object."""
+        self._errors = 0
+
+    def info(self, message: str) -> None:  # noqa: WPS110
+        """Utility trace info method."""
+        self._log_msg("INFO: {0}".format(message))
+
+    def warning(self, message: str) -> None:
+        """Utility trace warning method."""
+        self._log_msg("WARNING: {0}".format(message))
+
+    def warn(self, message: str) -> None:
+        """Utility trace warn method."""
+        self.warning(message)
+
+    def error(self, message: str) -> None:
+        """Utility trace error method."""
+        self._log_msg("ERROR: {0}".format(message))
+        self._errors += 1
+
+    def _dtrace(self, message: str) -> None:
+        self._log_msg("DEBUG: {0}".format(message))
+
+    def _itrace(self, message: str) -> None:
+        """Utility info trace method."""
+        self.info(message)
+
+    def _log_msg(self, message: str) -> None:
+        """Write message to stderr."""
+        print(message, file=sys.stderr)
+
+
+class Options(SimpleScribe):
     """A class to handle Options."""
 
     options: OptionsDict
-    scribe: Optional[Scribe]
 
     def __init__(self, opts: Optional[OptionsDict] = None) -> None:
         """Options constructor.
@@ -76,14 +123,14 @@ class Options(object):  # noqa: WPS214
         opts : OptionsDict Optional
             Options to merge with defaults
         """
+        super().__init__()
         self.options = {OPTIONS_DEBUG: 0, OPTIONS_VERBOSE: 0, OPTIONS_TEST: False}
-        self.errors = 0
-        self.scribe = None
         if opts is not None:  # pragma no cover ???
             self.options.update(opts)
-        if int(self.options[OPTIONS_DEBUG]) > 1:
-            print(  # noqa: WPS421
+        if int(self.options[OPTIONS_DEBUG]):
+            self.debug(
                 "created instance of class {0}".format(self.__class__.__name__),
+                2,
             )
 
     def isdebug(self) -> bool:
@@ -109,51 +156,11 @@ class Options(object):  # noqa: WPS214
     def debug(self, message: str, level: int = 1) -> None:
         """Utility debug method."""
         dl: int = int(self.options[OPTIONS_DEBUG])
-        if self.options[OPTIONS_TEST] or dl >= level:  # pragma no branch
+        if self.options[OPTIONS_TEST] or dl >= level:
             self._dtrace(message)
 
     def verbose(self, message: str, level: int = 1) -> None:
         """Utility verbose method."""
         vl: int = int(self.options[OPTIONS_VERBOSE])
-        if self.options[OPTIONS_TEST] or vl >= level:  # pragma no cover ???
+        if self.options[OPTIONS_TEST] or vl >= level:
             self._itrace(message)
-
-    def info(self, message: str) -> None:  # noqa: WPS110
-        """Utility trace info method."""
-        if self.scribe is None:
-            print("INFO: {0}".format(message))
-        else:
-            self.scribe.logger.info(message)
-
-    def warning(self, message: str) -> None:
-        """Utility trace warning method."""
-        if self.scribe is None:
-            print("WARNING: {0}".format(message))
-        else:
-            self.scribe.logger.warning(message)
-
-    def warn(self, message: str) -> None:
-        """Utility trace warn method."""
-        self.warning(message)
-
-    def error(self, message: str) -> None:
-        """Utility trace error method."""
-        if self.scribe is None:
-            print("ERROR: {0}".format(message))
-        else:
-            self.scribe.logger.error(message)
-        self.errors += 1
-
-    def _dtrace(self, message: str) -> None:
-        """Utility debug trace method."""
-        if self.scribe is None:
-            print("DEBUG: {0}".format(message))
-        else:
-            self.scribe.logger.debug(message)
-
-    def _itrace(self, message: str) -> None:
-        """Utility info trace method."""
-        if self.scribe is None:
-            print("INFO: {0}".format(message))
-        else:
-            self.scribe.logger.info(message)
