@@ -1,8 +1,7 @@
 """Top-level module for wtforglib Library."""
+
 import json
 from pathlib import Path
-from tempfile import TemporaryFile
-from typing import Optional, Tuple
 
 import yaml
 
@@ -11,35 +10,35 @@ from wtforglib.kinds import Fspec, StrAnyDict
 
 
 def load_yaml_file(
-    filenm: Fspec,
+    file_spec: Fspec,
     missing_ok: bool = True,
 ) -> StrAnyDict:
     """Loads a yaml file.
 
     Parameters
     ----------
-    filenm : Fspec
+    file_spec : Fspec
         The yaml file to load
     missing_ok : bool optional default True
-        When False and file does not exist raises FileNotFound exeception
+        When False and file does not exist raises FileNotFound exception
 
     Returns
     -------
     StrAnyDict
-        Representing contents of filenm
+        Representing contents of file_spec
     """
-    ypath = Path(filenm)
+    yaml_path = Path(file_spec)
     yaml_result = {}
-    if ypath.exists() and ypath.is_file():
-        with open(ypath, "r") as yfile:
-            yaml_result = yaml.safe_load(yfile)
+    if yaml_path.exists() and yaml_path.is_file():
+        with open(yaml_path, "r") as yaml_file:
+            yaml_result = yaml.safe_load(yaml_file)
     elif not missing_ok:  # pragma no cover
-        raise_filenotfound(ypath)
+        raise_filenotfound(yaml_path)
     return yaml_result
 
 
 def write_yaml_file(
-    filenm: Fspec,
+    file_spec: Fspec,
     src_data: StrAnyDict,
     encoding: str = "utf-8",
 ) -> bool:
@@ -47,7 +46,7 @@ def write_yaml_file(
 
     Parameters
     ----------
-    filenm : Fspec
+    file_spec : Fspec
         The yaml filename to write to
     src_data : StrAnyDict
         The data to write to a file
@@ -59,42 +58,42 @@ def write_yaml_file(
         bool
             True if file exists else False
     """
-    ypath = Path(filenm)
-    with open(ypath, "w") as outf:
-        yaml.dump(src_data, outf, encoding=encoding)
-    return ypath.exists()
+    yaml_path = Path(file_spec)
+    with open(yaml_path, "w") as out_file:
+        yaml.dump(src_data, out_file, encoding=encoding)
+    return yaml_path.exists()
 
 
 def load_json_file(
-    filenm: Fspec,
+    file_spec: Fspec,
     missing_ok: bool = True,
 ) -> StrAnyDict:
     """Loads a json file.
 
     Parameters
     ----------
-    filenm : Fspec
+    file_spec : Fspec
         The yaml file to load
     missing_ok : bool optional default True
-        When False and file does not exist raises FileNotFound exeception
+        When False and file does not exist raises FileNotFound exception
 
     Returns
     -------
     StrAnyDict
-        Representing contents of filenm
+        Representing contents of file_spec
     """
-    jpath = Path(filenm)
+    jpath = Path(file_spec)
     rtn: StrAnyDict = {}
     if jpath.exists() and jpath.is_file():
-        with open(jpath, "r") as jfile:
-            rtn = json.load(jfile)
+        with open(jpath, "r") as json_file:
+            rtn = json.load(json_file)
     elif not missing_ok:  # pragma no cover
         raise_filenotfound(jpath)
     return rtn
 
 
 def write_json_file(
-    filenm: Fspec,
+    file_spec: Fspec,
     src_data: StrAnyDict,
     indent: int = 2,
 ) -> bool:
@@ -102,7 +101,7 @@ def write_json_file(
 
     Parameters
     ----------
-    filenm : Fspec
+    file_spec : Fspec
         The json filename to write to
     src_data : StrAnyDict
         The data to write to a file
@@ -114,97 +113,7 @@ def write_json_file(
         bool
             True if file exists else False
     """
-    jpath = Path(filenm)
-    with open(jpath, "w") as outf:
-        outf.write(json.dumps(src_data, indent=2))
+    jpath = Path(file_spec)
+    with open(jpath, "w") as out_file:
+        out_file.write(json.dumps(src_data, indent=indent))
     return jpath.exists()
-
-
-def ensure_directory(target: Fspec, perm: int = 0o755) -> bool:
-    """Creates a directory if it doesn't exist.
-
-    Parameters
-    ----------
-    target : Fspec
-        The path to the target directory
-    perm : int Optional
-        The mode to use if creationg needed
-
-    Returns
-    -------
-    bool
-        True if directory exists
-
-    Raises
-    ------
-    NotADirectoryError
-        if target exists, but not a directory
-    """
-    dp = Path(target)
-    if dp.is_dir():
-        return True
-    if dp.exists():
-        raise NotADirectoryError(
-            'Pathname "{0}" is not a directory.'.format(str(target)),
-        )
-    dp.mkdir(mode=perm, parents=True)
-    return True
-
-
-def _verify_directory_write(dpath: Fspec) -> Optional[Exception]:
-    """Verify that the given directory is writable.
-
-    Parameters
-    ----------
-    dpath : Fspec
-        Pathlike object specifying the directory
-
-    Returns
-    -------
-    Optional[Exception]
-        The exception when caught
-    """
-    error: Optional[Exception] = None
-    try:
-        tfile = TemporaryFile(dir=dpath)
-        tfile.close()
-    except Exception as ex:
-        error = ex
-    return error
-
-
-def verify_directory(dspec: Fspec, ex: bool = False) -> Tuple[bool, str]:
-    """Verify that a directory exits and is writable.
-
-    Parameters
-    ----------
-    dspec : Fspec
-        Pathlike object specifying the directory
-    ex : bool, optional
-        When True exceptions are raised, by default False
-
-    Returns
-    -------
-    Tuple[bool, str]
-        When ex is False the status is returned
-
-    Raises
-    ------
-    error
-        NotADirectoryError, FileNotFoundError or other exception is raised on failure
-        when ex is True
-    """
-    dpath = Path(dspec)
-    dstr = str(dpath)
-    error: Optional[Exception] = None
-    if dpath.exists():
-        if dpath.is_dir():
-            error = _verify_directory_write(dpath)
-        else:
-            error = NotADirectoryError("'{0}' is not a directory".format(dstr))
-    else:
-        error = FileNotFoundError("Directory not found: {0}".format(dpath))
-    if error:
-        if ex:
-            raise error
-    return (error is None, str(error))
